Update api.py
Almost a complete rewrite. Lots of improvements:

-Refactoring and reformatting
-Better error handling
-More methods to import requests module
-Ability to automatically update
-Ability to disable update checks, or just ignore update checks for a specific version
-Detects whether it is being run on android (to make it easier to diagnose problems)
-Checks if Pixel Gun 3D is installed
-Automatically runs and prints results if executed directly, but not if imported as a module

I tested everything on my Windows 11 PC, but I did not test on android. For example, the android detection is untested.
I also credited myself as one of the authors and changed the version to 2.0.0. Feel free to change this if you don't like it.